# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FASE 2 API",
        "description": "FASE 2 Swagger OpenApi",
        "contact": {
            "email": "eaguilar.arrezola@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/purchases/paginate": {
            "post": {
                "tags": [
                    "purchases"
                ],
                "summary": "Get purchase per paginate",
                "operationId": "4d8833104408e8909e4247f437885d7e",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/purchases": {
            "get": {
                "tags": [
                    "purchases"
                ],
                "summary": "Get all purchases",
                "operationId": "6ff45401b4ef432e7fafaa91c8e9e513",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "purchases"
                ],
                "summary": "Add purchase",
                "operationId": "f1b6a3eaf182ed52b3c9d97ea2e8027a",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/purchases/{id}": {
            "get": {
                "tags": [
                    "purchases"
                ],
                "summary": "Get purchase",
                "operationId": "4aad3da11974dade35c9b2a902bc3e7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "purchases"
                ],
                "summary": "Update purchase",
                "operationId": "461bd92be49664b9dec01e83ea6235c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "purchases"
                ],
                "summary": "Delete purchase",
                "operationId": "68da2a0f195a3cf3d3f4482ddb775ad6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/purchases/pay/{id}": {
            "post": {
                "tags": [
                    "purchases"
                ],
                "summary": "Pay purchase",
                "operationId": "549f238baff08126b20f610dcb6d4c5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/purchases/cancel/{id}": {
            "post": {
                "tags": [
                    "purchases"
                ],
                "summary": "Cancel purchase",
                "operationId": "621ec5aaeac4ae262b8a5a6f565a0072",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}